requirements.txt
python-dotenv
flask 
Flask-SQLAlchemy
flasgger
firebase-admin
pandas
pyarrow
black
pytest 